exercise: 4.15-4.23

+++ part4/controllers/blogList.js
@@ -1,12 +1,39 @@
 const blogListRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blogList')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
+// 获取token
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+//     return authorization.substring(7)
+//   }
+//   return null
+// }
+
 blogListRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
-blogListRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogListRouter.post('/', userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
+  // const token = getTokenFrom(request)
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token missing or invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user._id
+  })
+
   if (request.body.title && request.body.url) {
     const saveBlog = await blog.save()
     response.json(saveBlog)
@@ -15,9 +42,21 @@ blogListRouter.post('/', async (request, response) => {
   }
 })
 
-blogListRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogListRouter.delete('/:id', userExtractor, async (request, response) => {
+  // 只有添加博客的用户才能删除博客
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({error: 'token missing or invalid'})
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({error: 'this not you blog'})
+  }
 })
 
 blogListRouter.put('/:id', async (request, response) => {
